+++ sregistry/client/delete.py
@@ -10,7 +10,7 @@ with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
 from sregistry.logger import bot
 
-def main(args, parser, subparser):
+def main(args, parser, extra):
 
     from sregistry.main import get_client