+++ sregistry/client/shell.py
@@ -9,7 +9,7 @@ with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 '''
 
 
-def main(args,parser,subparser):
+def main(args, parser, extra):
 
     lookup = { 'ipython': ipython,
                'python': python,